Change allowed diameter inputs
Now you can't set a stop diameter, it's fixed to 22μm; I'm fairly
certain that parameter values used in the distributions to generate the
tree depend on this value, so allowing it to change doesn't make much
sense. It doesn't really add anything, even if that weren't the case

Also, the initial diameter must be greater than or equal to 30μm.
Strictly speaking this only needs to be 22μm (the stop diamter) but
values below 30μm give bad looking trees, so might as well set the
minimum there.